Various improvements following industrial usage (#17)
Fix: handle timeout by polling socket instead of delegating to Linux kernel (kernel precision is too low for a realtime behavior @ 1 kHz) 
Update: error handler now takes the error reason in argument to let the user takes the right decision (a lost datagram is different from a non responsive slave)
Various cleanup and fixes (gcc warnings, ASAN and UBSAN feedback).

+++ src/Mailbox.cc
@@ -146,7 +146,7 @@ namespace kickcat
             }
             else
             {
-                header_->len += size;
+                header_->len += static_cast<uint16_t>(size);
                 std::memcpy(payload_, &size, sizeof(uint32_t));
                 std::memcpy(payload_ + sizeof(uint32_t), data, size);
             }
